Restore node 9.9.9, remove 9.10.1, 9.11.0, 9.11.1
CI had a bug where it removes the oldest version without preserving the
lastest version that was in the previous released buildpack. This is
undoing the recent mistakes that CI made. (In preparation for testing
the fix to CI, which will result in replaces 9.10.0 with 9.11.1).

Revert "Add node 9.11.1, remove node 9.10.1"
This reverts commit 8c30fa56eaeafd0be8438cde3060d96d49651fc8.

Revert "Add node 9.11.0, remove node 9.10.0"
This reverts commit a6e81fd5d0134ffc88991b30397f5e5c59fe42d5.

Revert "Add node 9.10.1, remove node 9.9.0"
This reverts commit 368ec9c0283e50398f0ebacf17f356982898f240.

Signed-off-by: Leah Hanson <lhanson@pivotal.io>