profile script: do not run exit command
It looks like lifecycle "source"s profile scripts, and therefore an
"exit" statement in the script stops the entire process starting the
app.

The error in the script only surfaces when $APPD_AGENT is set by the
user. The original PR submitted by appdynamics contributor[1] suggests
that the variable $APPD_AGENT is set by the user to signal a
Multibuildpack setup. This is when users explicitly use a dedicated
appdyanmics buildpack, and in that case, buildpacks like CF nodejs
buildpack do not set up appdyanmics and instead leaves it to the
appdyanmics buildpack. Looking for an example of such a buildpack, the
documentation of the proprietary tanzu appdynamics buildpack supports
this[2].

To reproduce the error:
cf push testapp -p fixtures/simple --no-start
cf set-env testapp APPD_AGENT "something"
cf restage testapp

1. https://github.com/cloudfoundry/nodejs-buildpack/pull/214
2. https://docs.vmware.com/en/AppDynamics-Application-Performance-Monitoring-for-VMware-Tanzu/services/appdynamics-application-performance-monitoring/using.html#nodejs-applications-multibuildpack-10